+++ Sources/Hoa3D/Tools_3D.h
@@ -25,10 +25,7 @@
 
 #ifdef _WINDOWS
 
-#include <ipps.h>
-#include <ippm.h>
-#include <mkl.h>
-#include <mkl_cblas.h>
+#include <gsl_cblas.h>
 
 #endif
 
@@ -109,7 +106,7 @@ namespace Hoa3D
 
 		if(l == m)
         {
-			return pow(-1.0f, (double)m) * pow(1. - x * x, 0.5 * m) * double_factorial(2. * m - 1);
+			return pow((double)-1.0, (double)m) * pow((double)(1. - x * x), (double)(0.5 * m)) * double_factorial(2. * m - 1);
 		}
 		else if(l == m + 1)
         {
